make current_x a size_t instead of an int
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2034 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/move.c
@@ -54,7 +54,7 @@ void do_home(void)
     size_t old_pww = placewewant;
 #ifndef NANO_SMALL
     if (ISSET(SMART_HOME)) {
-	int old_current_x = current_x;
+	size_t old_current_x = current_x;
 
 	for (current_x = 0; isblank(current->data[current_x]) &&
 		current->data[current_x] != '\0'; current_x++)
